+++ v1/scene/errors.go
@@ -25,6 +25,8 @@ const (
 	ErrPropAlreadyExist
 	ErrPropNotFound
 	ErrCleanupFailed
+	ErrNoSubscriber
+	ErrWaitedTooLong
 )
 
 var errorMessages = errors.Messages{
@@ -33,6 +35,8 @@ var errorMessages = errors.Messages{
 	ErrPropAlreadyExist: "prop %q already exist",
 	ErrPropNotFound:     "prop %q does not exist",
 	ErrCleanupFailed:    "cleanup of prop %q failed",
+	ErrNoSubscriber:     "no subscriber to signal %q",
+	ErrWaitedTooLong:    "waiting for signal %q timed out",
 }
 
 //--------------------
@@ -67,6 +71,18 @@ func IsPropNotFoundError(err error) bool {
 // failing of a prop error.
 func IsCleanupFailedError(err error) bool {
 	return errors.IsError(err, ErrCleanupFailed)
+}
+
+// IsNoSubscriberEroor returns true, if the error signals that
+// nobody waits for the passed signal.
+func IsNoSubscriberError(err error) bool {
+	return errors.IsError(err, ErrNoSubscriber)
+}
+
+// IsWaitedTooLongError returns true, if the error signals a
+// timeout when waiting for a signal.
+func IsWaitedTooLongError(err error) bool {
+	return errors.IsError(err, ErrWaitedTooLong)
 }
 
 // EOF