+++ v1/scene/scene_test.go
@@ -221,4 +221,139 @@ func TestAbort(t *testing.T) {
 	assert.ErrorMatch(err, "aborted")
 }
 
+// TestSignalNoTimeout tests the waiting for a signal without
+// a timeout.
+func TestSignalNoTimeout(t *testing.T) {
+	assert := asserts.NewTestingAssertion(t, false)
+	scn := scene.Start()
+
+	go func() {
+		err := scn.WaitSignal("foo")
+		assert.Nil(err)
+		err = scn.Store("foo-a", true)
+		assert.Nil(err)
+	}()
+	go func() {
+		err := scn.WaitSignal("foo")
+		assert.Nil(err)
+		err = scn.Store("foo-b", true)
+		assert.Nil(err)
+	}()
+
+	time.Sleep(100 * time.Millisecond)
+
+	err := scn.Signal("foo")
+	assert.Nil(err)
+
+	time.Sleep(250 * time.Millisecond)
+
+	fooA, err := scn.Fetch("foo-a")
+	assert.Nil(err)
+	assert.Equal(fooA, true)
+	fooB, err := scn.Fetch("foo-b")
+	assert.Nil(err)
+	assert.Equal(fooB, true)
+
+	err = scn.Stop()
+	assert.Nil(err)
+}
+
+// TestNoSignalDueToStop tests the waiting for a signal while
+// a scene is stopped.
+func TestNoSignalDueToStop(t *testing.T) {
+	assert := asserts.NewTestingAssertion(t, false)
+	scn := scene.Start()
+
+	go func() {
+		err := scn.WaitSignal("foo")
+		assert.True(scene.IsSceneEndedError(err))
+	}()
+	go func() {
+		err := scn.WaitSignal("foo")
+		assert.True(scene.IsSceneEndedError(err))
+	}()
+
+	time.Sleep(100 * time.Millisecond)
+
+	err := scn.Stop()
+	assert.Nil(err)
+}
+
+// TestSignalWithoutSubscriber tests the signalling with no subscriber
+func TestSignalWithoutSubscriber(t *testing.T) {
+	assert := asserts.NewTestingAssertion(t, false)
+	scn := scene.Start()
+
+	err := scn.Signal("foo")
+	assert.True(scene.IsNoSubscriberError(err))
+
+	err = scn.Stop()
+	assert.Nil(err)
+}
+
+// TestSubscriberRemoved tests that the subscribers are removed
+// after a signal.
+func TestSubscriberRemoved(t *testing.T) {
+	assert := asserts.NewTestingAssertion(t, false)
+	scn := scene.Start()
+
+	go func() {
+		err := scn.WaitSignal("foo")
+		assert.Nil(err)
+	}()
+	go func() {
+		err := scn.WaitSignal("foo")
+		assert.Nil(err)
+	}()
+
+	time.Sleep(100 * time.Millisecond)
+
+	err := scn.Signal("foo")
+	assert.Nil(err)
+
+	err = scn.Signal("foo")
+	assert.True(scene.IsNoSubscriberError(err))
+
+	err = scn.Stop()
+	assert.Nil(err)
+}
+
+// TestSignalTimeout tests the waiting for a signal with
+// a timeout.
+func TestSignalTimeout(t *testing.T) {
+	assert := asserts.NewTestingAssertion(t, false)
+	scn := scene.Start()
+
+	go func() {
+		err := scn.WaitSignal("foo")
+		assert.Nil(err)
+		err = scn.Store("foo-a", true)
+		assert.Nil(err)
+	}()
+	go func() {
+		err := scn.WaitSignalLimited("foo", 50*time.Millisecond)
+		assert.True(scene.IsWaitedTooLongError(err))
+		err = scn.Store("foo-b", true)
+		assert.Nil(err)
+	}()
+
+	time.Sleep(100 * time.Millisecond)
+
+	err := scn.Signal("foo")
+	assert.Nil(err)
+
+	fooA, err := scn.Fetch("foo-a")
+	assert.Nil(err)
+	assert.Equal(fooA, true)
+	fooB, err := scn.Fetch("foo-b")
+	assert.Nil(err)
+	assert.Equal(fooB, true)
+
+	err = scn.Signal("foo")
+	assert.True(scene.IsNoSubscriberError(err))
+
+	err = scn.Stop()
+	assert.Nil(err)
+}
+
 // EOF